Major edits: expanded introduction sections
I re-worked the introduction sections: Positive-Only Sparse Coding,
Mathematical Interpretation, and Neural Interpretation. They move more
slowly and repeat themselves. I also refactored some of the code so
that, e.g., the error function is always referred to as such.

You'll notice some rather ebullient use of Markdown, a couple of
discussion-type questions, and a few tangents, mostly relegated to
footnotes. If you're unsure of the purpose of any of these pieces,
just let me know so we can make sure we're all on the same page as
teachers.

I also switched all the color maps for displaying the digits to
'gray_r' instead of 'gray', so that the digits would be black on white.